[feat] Revamp blob data loading
