+++ pkgs/aimstack/asp/models/objects/audio.py
@@ -25,7 +25,7 @@ class Audio(Object):
     """
 
     AIM_NAME = 'aim.Audio'
-    RESOLVE_BLOBS = True
+    RESOLVE_BLOBS = False
 
     # supported audio formats
     UNKNOWN = ''
