+++ src/aimcore/web/ui/src/modules/core/engine/blob-uri-system/index.ts
@@ -1,8 +1,6 @@
 import _ from 'lodash-es';
 
-import { createBlobsRequest } from 'modules/core/api/runsApi';
-
-import { SequenceType } from 'types/core/enums';
+import { createBlobsRequest } from 'modules/core/api/dataFetchApi';
 
 import { parseStream } from 'utils/encoder/streamEncoding';
 import arrayBufferToBase64 from 'utils/arrayBufferToBase64';
@@ -22,15 +20,13 @@ export interface IBlobURISystemEngine {
 
 const BATCH_SEND_DELAY = 1000;
 
-function createBlobURISystemEngine(
-  sequenceType: SequenceType,
-): IBlobURISystemEngine {
+function createBlobURISystemEngine(): IBlobURISystemEngine {
   const blobsData: Record<string, string> = {};
   const blobsSubscriptions: Record<string, Callback[]> = {};
   let blobUriQueue: string[] = [];
   let timeoutID: number | null = null;
 
-  const request = createBlobsRequest(sequenceType);
+  const request = createBlobsRequest();
 
   /**
    * Function to fire an event
